+++ tests/providers/azure/services/app/app_function_application_insights_enabled/app_function_application_insights_enabled_test.py
@@ -87,6 +87,7 @@ class Test_app_function_application_insights_enabled:
                         public_access=False,
                         vnet_subnet_id=None,
                         ftps_state="AllAllowed",
+                        resource_group_name="resource_group_name",
                     )
                 }
             }
@@ -137,6 +138,7 @@ class Test_app_function_application_insights_enabled:
                         public_access=False,
                         vnet_subnet_id=None,
                         ftps_state="AllAllowed",
+                        resource_group_name="resource_group_name",
                     )
                 }
             }
@@ -187,6 +189,7 @@ class Test_app_function_application_insights_enabled:
                         public_access=False,
                         vnet_subnet_id=None,
                         ftps_state="AllAllowed",
+                        resource_group_name="resource_group_name",
                     )
                 }
             }
@@ -237,6 +240,7 @@ class Test_app_function_application_insights_enabled:
                         public_access=False,
                         vnet_subnet_id=None,
                         ftps_state="AllAllowed",
+                        resource_group_name="resource_group_name",
                     )
                 }
             }